I'm refactoring the domain spam detection system to improve the task of preventing users from spamming the system.
The new system is more simplified and streamlined. It allows users to 
pick whether they wish to white or blacklist a domain or IP, stores this
information to a database and then appropriately detects whether the 
domain was blacklisted.

Some changes:

1. Removed the auto extend feature. While an interesting concept, the 
idea of the system automatically blacklisting domains or IPs in a 
extending pattern seemed like a great idea. But in the end it just
overengineers a solution to a problem that's too uncommon.

2. Removed the option to "blacklist + subdomains". The system now asumes
that a blacklisted domain implies all subdomains are blacklisted. This
can be overridden by whitelisting the appropriate subdomains.

TODO:

1. IP block based blocking. When an IP does end in 0s... It's actually
a subnet and implies that all IP in the range should be blocked.