Touched up the user profiles. I am currently working on an implementation to micro-serviced webhooks that does not suck when trying to work with it.
The service needs to be able to access the list of apps installed
on the server to allow an administator using the new hook app to
connect to applications from the list.

It would also be recommended to have a way that allows applications
integrated with the server to provide a list of hooks that they
support. This way a user could potentially manage the hooks from
the UI. Probably, the smartest way to achieve this is by letting
the app upload a JSON file with supported hooks, captions and
descriptions (including hierarchies)

On the other hand, apps should be able to create hooks themselves
by registering them or checking whether the hooks they need already
do exist.

Applications should be able to request "transliteration".
Sometimes the payload from one app does not fit the API of the
receiving app, when the vendor of one of the apps is not ourselves,
this could could become a problem if the server does not allow
translating the formats slightly.