Email editing form moved to vue (#50)
* Email editing form moved to vue

By moving the email editing form to tailwind and vue we're adding
a series of ux improvements. The system can now verify the email is not
the same, nor an invalid email before it is relayed to the server. This
reduces the issues users have when they enter the same email again and
they're warned that the email is already in use.