Record when a suspension was created
This now records when a suspension was recorded and therefore allows
admins to see how long a suspension lasted or is lasting.

Note that this has a side effect. When a very old suspension is
manipulated, the system will record the current time into it. This means
that lifting an already existing suspension may lead to the application
showing that it was created the moment the suspension was lifted.