Replace IPStack with Cloudflare Geo
Cloudflare includes a Geolocation service which will automatically add
the user's country and city to the headers. The service is included in
CFs base plan, and while it produces some vendor lock-in it should do a
better job at not running out of requests than IP stack has.

In future revisions I would love to replace this vendor specific
integration with a generic interface for geolocation. Also, note that
the webserver needs to be configured to ignore any requests not coming
from cloudflare for this system to be secure.