Adds support for basic discovery
This should make it easy for openID style clients to log into the
server and authenticate tokens. Once the public keys can be properly
cached by the clients it will significantly reduce the load on the
identity server since it will no longer need to verify each token on
each request.