Added basic VPN detection
If a session is initiated from a VPN, the server will record the fact
that the timezone on the client and the country they're allegedly
visiting from do not match.

Please note that automatic enforcement of VPN bans is not intended. VPNs
do have legitimate use cases, but it may be a red flag that compounds
with other red flags to show that a user is being shifty.