Introduces oAuth 2.0 support (#58)
* Introduces oAuth 2.0 support

This introduces several mechanisms needed for proper oAuth support.

1. Tokens are incinerated once they're expired.
2. Refresh tokens are incinerated
3. Sessions are properly incinerated
4. Applications are notified when a user terminates a session
5. Tokens are terminated when the session is terminated

* oAuth dialog is now more appealing

The old dialog has quite a few dated CSS references in there

* Adds support for basic discovery

This should make it easy for openID style clients to log into the
server and authenticate tokens. Once the public keys can be properly
cached by the clients it will significantly reduce the load on the
identity server since it will no longer need to verify each token on
each request.

* Display the new sessions on the account index

This allows a user to terminate all the tokens that depend on a single
session in one fell swoop.

* Added defer of session incineration

PHPAS should automatically schedule tasks to incinerate sessions once
the user logs out.

* Finishing touches to the oauth2 dialog

The oauth2 dialog now looks properly implemented. I also sprinkled a
bunch of todos in betweeen the CSS so I can remember that they slowly
need to go

* Basic token rotation

This is not yet invalidating sessions when the user trips a fence, since
we currently don't have any applications that would be affected by the
change. But it is definitely a priority to keep users safe

* Collections are not supported by query yet

Needed to convert that to an array first.

* Redis queues do not need locks.

Redis is atomic. It won't need locks to synchronize the different
workers picking up the same task.

* Superfluous if checks removed in User::activate

The user controller had superfluous checks to make sure the token is set
in a context where it was clear that the token was empty.

* Add token incineration tasks.

These allow the application to defer incineration of the tokens,
reducing stress on the database at the time of handling. Instead, we can
have a more diligent incinerator task running in the background and
notifying other resources when needed.

* When ending a session end the depending tokens

Also incinerate them to recover some disk space.

* Keys are bound to the server.

Since the server does the signing and not the applications themselves,
we do not need scoped keys for each app. We can just define a certain
key as canonical.

* Added basic VPN detection

If a session is initiated from a VPN, the server will record the fact
that the timezone on the client and the country they're allegedly
visiting from do not match.

Please note that automatic enforcement of VPN bans is not intended. VPNs
do have legitimate use cases, but it may be a red flag that compounds
with other red flags to show that a user is being shifty.

* Notify apps of ended sessions.

This allows PHPAuthserver to send a notification to an application that
depends on it for managing it's sessions.

* Updated logout

The logout should also incinerate the session appropriately. Reducing
friction when logging out of an application.

* Added todo regarding session sanity checks

The session is sanity checked on each request. Ensuring that the db
mirror of the session exists when the user requests any page. This seems
way more complicated and convoluted than it should.

* Style checked.