Fixes #80: SQL Error when username doesn't exist
When a user tries to log into an account without a username (either
because it has a typo or because they're using their email), the system
will issue an SQL error.

Also taking the opportunity to trim the username or email (since
autocompletion on Android and iOS loves to append a stray space when
filling in forms for you)

Thanks to @docmew for reporting the issue. :)