Session: Check session is started w/o session_id()
The session_id() function is a bit of a tricky beast, since PHP will
only have it return a value AFTER the session was started. This does not
imply that the user has no session.

For example, the session cookie might be set, but the session_id() will
return null because we have not yet read the session... We do want
(though) to read the session ID to check whether there's data for the
given session in the DB before opening the actual file on the server.